+++ NEWS.md
@@ -2,6 +2,8 @@ Listing news on any major breaking changes in DFG.  For regular changes, see int
 
 # v0.27
 - `delete` returns number of nodes deleted and no longer the object that was deleted.
+- Deprecate `updateVariable!` for `mergeVariable!`, note `merege` returns number of nodes updated/added.
+- Deprecate `updateFactor!` for `mergeFactor!`, note `merege` returns number of nodes updated/added.
 
 # v0.26
 - Graph structure plotting now uses GraphMakie.jl instead of GraphPlot.jl. Update by replacing `using GraphPlot` with `using GraphMakie`.